Fix haste formula, and add tests
