+++ packages/xivmath/src/test/gcd_comparison_tests.ts
@@ -1,12 +1,81 @@
 import {getLevelStats} from "../xivconstants";
 import {
+    fl,
     sksToGcd,
-    sksToGcd_etroOriginal,
-    sksToGcd_etroSimplified,
-    sksToGcd_newRounding,
-    sksToGcd_simplified
 } from "../xivmath";
 import {assert} from "chai";
+import {LevelStats} from "../geartypes";
+
+export function sksToGcd_simplified(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
+    return fl((100 - haste) * ((baseGcd * (1000 - commonPart1)) / 1000)) / 100;
+}
+
+export function sksToGcd_newRounding(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
+    return fl((100 - haste) * fl((baseGcd * (1000 - commonPart1)) / 1000)) / 100;
+}
+
+// Not working
+export function sksToGcd_etroOriginal(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
+    const arrow = 0,
+        feyWind = 0,
+        selfBuff2 = 0,
+        RoF = 100,
+        umbralAstral3 = 100;
+    return (
+        fl(
+            (fl(
+                    (fl(
+                            (fl(
+                                    ((1000 -
+                                            fl(
+                                                (130 * (sks - levelStats.baseSubStat)) / levelStats.levelDiv
+                                            )) *
+                                        baseGcd) /
+                                    1000
+                                ) *
+                                fl(
+                                    ((fl(
+                                                (fl(((100 - arrow) * (100 - haste)) / 100) *
+                                                    (100 - 0)) /
+                                                100
+                                            ) -
+                                            feyWind) *
+                                        (selfBuff2 - 100)) /
+                                    100
+                                )) /
+                            -100
+                        ) *
+                        RoF) /
+                    1000
+                ) *
+                umbralAstral3) /
+            100
+        ) / 100 // Added to taken func. Converts from MS to S.
+    );
+}
+
+// Not working
+export function sksToGcd_etroSimplified(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
+    return Math.floor(
+        (Math.floor(
+            (Math.floor(
+                (Math.floor((baseGcd * (1000 - commonPart1) / 1000)) *
+                    Math.floor((100 - haste) * -1)) / -100) * 100) / 1000) * 100) / 100) / 100;
+}
+
+export function sksToGcd_makarOriginal(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    return fl(fl(fl(fl((1000 - fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv)) * (baseGcd * 1000) / 1000) * fl((fl(fl((100 - 0) * (100 - 0) / 100) * (100 - haste) / 100) - 0) * (0 - 100) / 100) / -100) * 100 / 1000) * 100 / 100) / 100;
+}
+
+export function sksToGcd_makarSimplified(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
+    return fl(fl(fl((1000 - fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv)) * baseGcd) * (100 - haste)) / 1000) / 100;
+}
+
+
 
 type Result = {
     sks: number,
@@ -15,7 +84,9 @@ type Result = {
     simplified: number,
     newRounding: number,
     etroOriginal: number,
-    etroSimplified: number
+    etroSimplified: number,
+    makarOriginal: number,
+    makarSimplified: number
 }
 describe('sks calc', () => {
     it('comparison', () => {
@@ -24,7 +95,7 @@ describe('sks calc', () => {
         const baseSks = levelStats.baseSubStat;
         const baseGcd = 2.5;
         for (let sks = baseSks; sks < 1000; sks++) {
-            for (let haste = 0; haste < 20; haste++) {
+            for (let haste = 0; haste < 25; haste++) {
                 results.push({
                     sks: sks,
                     haste: haste,
@@ -33,14 +104,21 @@ describe('sks calc', () => {
                     newRounding: sksToGcd_newRounding(baseGcd, levelStats, sks, haste),
                     etroOriginal: sksToGcd_etroOriginal(baseGcd, levelStats, sks, haste),
                     etroSimplified: sksToGcd_etroSimplified(baseGcd, levelStats, sks, haste),
+                    makarOriginal: sksToGcd_makarOriginal(baseGcd, levelStats, sks, haste),
+                    makarSimplified: sksToGcd_makarSimplified(baseGcd, levelStats, sks, haste),
                 })
             }
         }
         const bad = results.filter(resultRow => {
-            return resultRow.original !== resultRow.simplified
-                || resultRow.original !== resultRow.newRounding
+            return (
+                // false
+                // || resultRow.original !== resultRow.simplified
+                // || resultRow.original !== resultRow.newRounding
                 // || resultRow.original !== resultRow.etroOriginal
                 // || resultRow.original !== resultRow.etroSimplified
+                resultRow.makarSimplified !== resultRow.makarOriginal
+                || resultRow.makarSimplified !== resultRow.original
+            )
         });
         bad.forEach(resultRow => {
             console.log(JSON.stringify(resultRow));
