+++ packages/xivmath/src/xivmath.ts
@@ -59,66 +59,7 @@ function flp(places: number, input: number) {
  * @param haste The haste value, e.g. 15 for 15% haste, etc.
  */
 export function sksToGcd(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
-    return fl((100 - haste) * (baseGcd * 1000 * (1000 - fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv)) / 1000 / 1000)) / 100;
-}
-
-export function sksToGcd_simplified(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
-    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
-    return fl((100 - haste) * ((baseGcd * (1000 - commonPart1)) / 1000)) / 100;
-}
-
-export function sksToGcd_newRounding(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
-    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
-    return fl((100 - haste) * fl((baseGcd * (1000 - commonPart1)) / 1000)) / 100;
-}
-
-export function sksToGcd_etroOriginal(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
-    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
-    const arrow = 0,
-        feyWind = 0,
-        selfBuff2 = 0,
-        RoF = 100,
-        umbralAstral3 = 100;
-    return (
-        fl(
-            (fl(
-                    (fl(
-                            (fl(
-                                    ((1000 -
-                                            fl(
-                                                (130 * (sks - levelStats.baseSubStat)) / levelStats.levelDiv
-                                            )) *
-                                        baseGcd) /
-                                    1000
-                                ) *
-                                fl(
-                                    ((fl(
-                                                (fl(((100 - arrow) * (100 - haste)) / 100) *
-                                                    (100 - 0)) /
-                                                100
-                                            ) -
-                                            feyWind) *
-                                        (selfBuff2 - 100)) /
-                                    100
-                                )) /
-                            -100
-                        ) *
-                        RoF) /
-                    1000
-                ) *
-                umbralAstral3) /
-            100
-        ) / 100 // Added to taken func. Converts from MS to S.
-    );
-}
-
-export function sksToGcd_etroSimplified(baseGcd: number, levelStats: LevelStats, sks: number, haste = 0): number {
-    const commonPart1 = fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv);
-    return Math.floor(
-        (Math.floor(
-            (Math.floor(
-                (Math.floor((baseGcd * (1000 - commonPart1) / 1000)) *
-                    Math.floor((100 - haste) * -1)) / -100) * 100) / 1000) * 100) / 100) / 100;
+    return fl((fl((1000 - fl(130 * (sks - levelStats.baseSubStat) / levelStats.levelDiv)) * baseGcd) * (100 - haste)) / 1000) / 100;
 }
 
 /**
@@ -130,7 +71,7 @@ export function sksToGcd_etroSimplified(baseGcd: number, levelStats: LevelStats,
  * @param haste The haste value, e.g. 15 for 15% haste, etc.
  */
 export function spsToGcd(baseGcd: number, levelStats: LevelStats, sps: number, haste = 0): number {
-    return fl((100 - haste) * (baseGcd * 1000 * (1000 - fl(130 * (sps - levelStats.baseSubStat) / levelStats.levelDiv)) / 1000 / 1000)) / 100;
+    return fl((fl((1000 - fl(130 * (sps - levelStats.baseSubStat) / levelStats.levelDiv)) * baseGcd) * (100 - haste)) / 1000) / 100;
 }
 
 /**
@@ -492,5 +433,5 @@ export function vitToHp(levelStats: LevelStats, jobStats: JobData, vitality: num
     // TODO make this work without ts-ignore
     // @ts-expect-error - can't figure out type def
     const hpMod = levelStats.hpScalar[jobStats.role] ?? levelStats.hpScalar.other;
-    return fl(levelStats.hp * jobStats.jobStatMultipliers.hp / 100) + fl( (vitality - levelStats.baseMainStat) * hpMod);
+    return fl(levelStats.hp * jobStats.jobStatMultipliers.hp / 100) + fl((vitality - levelStats.baseMainStat) * hpMod);
 }
